requirements.txt
flask
deepface
tf-keras
opencv-python
numpy
flask-socketio
python-dotenv
authlib
flask-login
requests
librosa
soundfile
